test(config): the run half moves out too, and the fixtures get one home
The file-length ratchet, run against origin/main with the lane's files
staged, failed: tests/test_config_read_once.py reached 812 lines, and a
file this lane ADDS is failed for crossing 800 rather than grandfathered.
Measured rather than eyeballed, and it is why this is a second split
after the guard half already moved.

Three files now, split by what they drive rather than by size alone:
test_config_guard.py walks source, test_config_read_once.py drives the
real _phase_verify, and test_run_envelope.py drives run_factory. Each
docstring names the other two.

tests/helpers/run_config.py holds what both behaviour halves need: the
three kstrl.toml bodies, the mid-run edit, the git repo the diff phase
needs, and drive_run/empty_run. One home rather than two copies, because
two copies of a fixture drift into two different measurements.

Ratchets after the split, staged in a detached worktree at origin/main
so they compare against main rather than against the branch's own HEAD:
file length exit 0 with five already-oversize files reported and nothing
crossing; cyclomatic exit 0, one IMPROVED (load_or_report 2 -> 1) and no
NEW or REGRESSED function.
